+++ lib/src/models/environment_type.dart
@@ -2,7 +2,7 @@
 enum EnvironmentType { development, test, production }
 
 /// An extension method that takes the [EnvironmentType] enum, convert it to [String] and returns the last part of the enum.
-extension EnvironmentTypeToShortString on EnvironmentType {
+extension EnvironmentTypeToShortString on Enum {
   String toShortString() {
     return toString().split('.').last;
   }