+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+ - update docs
+ - add method initWithCustomType - for use custom environment type from entrypoint
+ - add method initFromJsonWithCustomType - for use custom environment type from json
+
 ## 1.0.1
  - environment variables initialization refactor